[Fix][P0][命名] UserName → Username 命名统一化修复
• 修复AppDbContextModelSnapshot.cs中列名不一致问题
  - 统一数据库快照文件中的列名为 "Username"
  - 确保与AppDbContext.cs和实体定义完全一致
• 修复MedicalCaseController中类型转换问题
  - UpdateMedicalCaseStatusDto.Status枚举类型转换为int
  - 保持与服务层UpdateStatus方法签名兼容
• 确认其他UserName使用均为正当API调用
  - Environment.UserName: .NET框架正确用法
  - f.Internet.UserName(): Bogus测试库正确用法
  - 架构测试中的禁用字段检查: 架构合规正确用法
• 解决TECH_DEBT_BACKLOG.md P0-9项技术债务

+++ src/Server/Services/LYBT.WebAPI/Controllers/MedicalCaseController.cs
@@ -275,7 +275,7 @@ public class MedicalCaseController : BaseApiController
                 return validation;
             }
 
-            var result = await _medicalCaseService.UpdateStatus(id, dto.Status);
+            var result = await _medicalCaseService.UpdateStatus(id, (int)dto.Status);
             return HandleBoolServiceResult(result, "状态更新成功");
         }
         catch (Exception ex)
